+++ Test/Unit/Model/Collector/StorefrontProbeCollectorTest.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace MageWatch\Agent\Test\Unit\Model\Collector;
 
 use MageWatch\Agent\Model\Collector\StorefrontProbeCollector;
+use MageWatch\Agent\Model\MagentoErrorPageDetector;
 use Magento\Store\Api\Data\StoreInterface;
 use Magento\Store\Model\StoreManagerInterface;
-use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
 class StorefrontProbeCollectorTest extends TestCase
@@ -20,19 +20,20 @@ class StorefrontProbeCollectorTest extends TestCase
         $storeManager = $this->createMock(StoreManagerInterface::class);
         $storeManager->method('getStore')->willReturn($store);
 
-        $collector = new StorefrontProbeCollector($storeManager);
+        $collector = new StorefrontProbeCollector($storeManager, new MagentoErrorPageDetector);
         $result = $collector->collect();
 
         $this->assertArrayHasKey('storefront_probe', $result);
         $this->assertSame('none', $result['storefront_probe']['probe_method']);
         $this->assertFalse($result['storefront_probe']['homepage_ok']);
         $this->assertFalse($result['storefront_probe']['checkout_ok']);
+        $this->assertFalse($result['storefront_probe']['homepage_magento_error']);
     }
 
     public function testCollectorCode(): void
     {
         $storeManager = $this->createMock(StoreManagerInterface::class);
-        $collector = new StorefrontProbeCollector($storeManager);
+        $collector = new StorefrontProbeCollector($storeManager, new MagentoErrorPageDetector);
 
         $this->assertSame('storefront_probe', $collector->getCode());
     }